repair 0

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -11,7 +11,7 @@ unsigned int binary_to_uint(const char *b)
 	int len;
 	int i;
 	int array[1024];
-	int sum = 0;
+	unsigned long int sum = 0;
 
 	if (b == NULL)
 	{
